% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{add.kegg.edge.mut}
\alias{add.kegg.edge.mut}
\title{Add edge to GraphNEL graph}
\usage{
add.kegg.edge.mut(entry1, entry2, subtype, g)
}
\arguments{
\item{entry1}{interactor 1}

\item{entry2}{interactor 2}

\item{subtype}{interaction subypte}

\item{g}{GraphNEL object}
}
\value{
GraphNEL object containing modified pathway
}
\description{
Add edge to GraphNEL graph
}
