% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{add.kegg.edge}
\alias{add.kegg.edge}
\title{Add edge to GraphNEL graph}
\usage{
add.kegg.edge(entry1, entry2, type, subtype, edge.attrs, g)
}
\arguments{
\item{entry1}{interactor 1}

\item{entry2}{interactor 2}

\item{type}{KEGG edge relation type (eg. PPrel - protein protein relation, GErel- gene expression relation)}

\item{subtype}{interaction subypte}

\item{edge.attrs}{edge attributes}

\item{g}{GraphNEL object}
}
\value{
GraphNEL object containing modified pathway
}
\description{
Add edge to GraphNEL graph
}
