% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{addEdgeSafe}
\alias{addEdgeSafe}
\title{Add an edge between two nodes if no such edge exists,
and if no reverse edge exists}
\usage{
addEdgeSafe(sourceNode, targetNode, g)
}
\arguments{
\item{sourceNode}{The source node}

\item{targetNode}{The target node}

\item{g}{The graph of class graphNEL}
}
\value{
g The modified graph with added (or not) edge
}
\description{
Add an edge between two nodes if no such edge exists,
and if no reverse edge exists
}
