% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathnet.functions.R
\name{calc_node_partial_influences}
\alias{calc_node_partial_influences}
\title{Returns ordered list of the nodes by their influence on the signal of specified nodes.}
\usage{
calc_node_partial_influences(
  pathway,
  influence_node,
  influence_direction = "any"
)
}
\arguments{
\item{pathway}{pathway with mapped gene expression data}

\item{influence_node}{sigle id of node or vector of node ids based on which partial influencse will be calculated.}

\item{influence_direction}{in which direction node affects the signal of target node. possible values c("+", "-", "any"). Default value is "any"}
}
\description{
Returns ordered list of the nodes by their influence on the signal of specified nodes.
}
