% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{correctEdgeDirections}
\alias{correctEdgeDirections}
\title{Guess wrong directed binding interactions and reverse them}
\usage{
correctEdgeDirections(g, edge.attrs)
}
\arguments{
\item{g}{The graph object of graphNEL class}

\item{edge.attrs}{The list of edge attributes in the graph}
}
\value{
g The modified graph
}
\description{
Guess wrong directed binding interactions and reverse them
}
