% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_denorm.R
\name{data_norm}
\alias{data_norm}
\title{Normalize the data}
\usage{
data_norm(data_in)
}
\arguments{
\item{data_in}{is a series of data}
}
\value{
the normalized data series
}
\description{
takes a series of data and normalize it
}
\examples{
#iris data as input
x <- na.omit(iris[2])
y <- data_norm(x)
y
}

