% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_psf_functions.R
\name{determine.input.nodes}
\alias{determine.input.nodes}
\title{Returns vector of node ids which do not have incoming edges but only outgoing.}
\usage{
determine.input.nodes(pathway)
}
\arguments{
\item{pathway}{pathway list object generated with package.}
}
\description{
Returns vector of node ids which do not have incoming edges but only outgoing.
}
