% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathnet.functions.R
\name{determine.sink.nodes}
\alias{determine.sink.nodes}
\title{Detects terminal nodes of the pathway}
\usage{
determine.sink.nodes(pathway)
}
\arguments{
\item{pathway}{pathway}
}
\description{
Detects terminal nodes of the pathway
}
