% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_psf_functions.R
\name{df_to_graphnel}
\alias{df_to_graphnel}
\title{Converts 2 data frames(node_table, edge_table) to graphNEL object}
\usage{
df_to_graphnel(node_table, edge_table)
}
\arguments{
\item{node_table}{node data frame with specified columns. Mandatory columns: label, component_id_s, type. Non mandatory columns: psf_function, x, y.}

\item{edge_table}{edge data frame with specified columns. Mandatory columns: from, to, type. Non mandatory columns subtype, state, weight.}
}
\description{
Converts 2 data frames(node_table, edge_table) to graphNEL object
}
