% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{download.KGML}
\alias{download.KGML}
\title{Download a KEGG pathway from the web}
\usage{
download.KGML(id, dir)
}
\arguments{
\item{id}{KEGG identifier of the pathway}

\item{dir}{The directory where the downloaded file should be saved}
}
\value{
The filepath of the downloaded pathway or NULL the pathway was not downloaded
}
\description{
Download a KEGG pathway from the web
}
