% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psf.graph.functions.R
\name{edge.exists}
\alias{edge.exists}
\title{Check if the edge exists in the graph}
\usage{
edge.exists(g, from, to)
}
\arguments{
\item{g}{graphNEl graph}

\item{from}{parent node id}

\item{to}{child node id}
}
\description{
Check if the edge exists in the graph
}
