% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{edge_data_frame_from_graph}
\alias{edge_data_frame_from_graph}
\title{Export data frame from graphNEL graph for edge data and its attributes}
\usage{
edge_data_frame_from_graph(g)
}
\arguments{
\item{g}{graphNEL graph}
}
\description{
Export data frame from graphNEL graph for edge data and its attributes
}
\author{
Siras Hakobyan
}
