% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{generate.kegg.collection.from.kgml}
\alias{generate.kegg.collection.from.kgml}
\title{Generate kegg collection from kgml files}
\usage{
generate.kegg.collection.from.kgml(kgml.files, sink.nodes = T)
}
\arguments{
\item{kgml.files}{chrachter vector of kgml paths}

\item{sink.nodes}{determin sink nodes, logical TRUE or FALSE}
}
\description{
Generate kegg collection from kgml files
}
