% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{generate.kegg.collection}
\alias{generate.kegg.collection}
\title{Download kegg pathways of provided pathway id list from keggrest and generate kegg collection}
\usage{
generate.kegg.collection(pathway.id.list, out.dir, sink.nodes = T)
}
\arguments{
\item{pathway.id.list}{charachter vector of pathway ids (hsa04151)}

\item{out.dir}{path to the directory where kgmls will be downloded}

\item{sink.nodes}{determin sink nodes, logical TRUE or FALSE}
}
\description{
Download kegg pathways of provided pathway id list from keggrest and generate kegg collection
}
