% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psf.graph.functions.R
\name{get.edge.type}
\alias{get.edge.type}
\title{Returns the general edge type (either activation or inhibition)}
\usage{
get.edge.type(g, from, to)
}
\arguments{
\item{g}{graphNEl graph}

\item{from}{parent node id}

\item{to}{child node id}
}
\description{
Returns the general edge type (either activation or inhibition)
}
