% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{get.pathway.attrs}
\alias{get.pathway.attrs}
\title{Get pathway general attributes}
\usage{
get.pathway.attrs(kgml)
}
\arguments{
\item{kgml}{path to kgml file}
}
\description{
Get pathway general attributes
}
