% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psf.structures.node-getValue-method.R
\name{getValue,Node-method}
\alias{getValue,Node-method}
\title{Get the value of a node}
\usage{
\S4method{getValue}{Node}(nodeObj)
}
\arguments{
\item{nodeObj}{The Node object}
}
\value{
the value of the Node
}
\description{
Get the value of a node
}
