% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_psf_functions.R
\name{graphnel_to_df}
\alias{graphnel_to_df}
\title{Converts graphNEL object to 2 data frames(node_table, edge_table)}
\usage{
graphnel_to_df(pathway, extended = FALSE)
}
\arguments{
\item{pathway}{pathway list object generated with package.}

\item{extended}{set to TRUE to get more detailed data frames with graphical information. Default value is FALSE.}
}
\description{
Converts graphNEL object to 2 data frames(node_table, edge_table)
}
