% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_psf_functions.R
\name{interactive_spatial_plot}
\alias{interactive_spatial_plot}
\title{Renders interactive plot of spatial tissue slice.}
\usage{
interactive_spatial_plot(
  seurat_object,
  feature = NULL,
  group = "seurat_clusters",
  log_norm = T
)
}
\arguments{
\item{seurat_object}{Seurat object of spatial transcriptomics dataset.}

\item{feature}{If feature is provided spots of the tissue slice will be color coded by feature values.}

\item{group}{group variable from meta.data object based on which spots will be colored. Default values is "seurat_clusters".}

\item{log_norm}{log transform spot values before color mapping. Default value is TRUE.}
}
\description{
Renders interactive plot of spatial tissue slice.
}
