% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{isReverseDirection}
\alias{isReverseDirection}
\title{This function predicts if the edge diractions are wrong based on graphical position of the KEGG nodes}
\usage{
isReverseDirection(snode, tnode, g)
}
\arguments{
\item{snode}{parent node id}

\item{tnode}{child node id}

\item{g}{a graphNEL graph}
}
\value{
TRUE or FALSE
}
\description{
This function predicts if the edge diractions are wrong based on graphical position of the KEGG nodes
}
