% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psf.graph.functions.R
\name{out.edges}
\alias{out.edges}
\title{provide outgoing edges of the specified node}
\usage{
out.edges(g, from)
}
\arguments{
\item{g}{graphNEL graph}

\item{from}{node id}
}
\description{
provide outgoing edges of the specified node
}
