% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{parse.KGML}
\alias{parse.KGML}
\title{Parse KGML to a graph object}
\usage{
parse.KGML(kgml)
}
\arguments{
\item{kgml}{The KGML file path}
}
\value{
GraphNEL object containing the parsed pathway
}
\description{
Parse KGML to a graph object
}
