% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{plot.pathway}
\alias{plot.pathway}
\title{Plots the pathway with KEGG layout based on x and y coordinates taken from KGML file.}
\usage{
\method{plot}{pathway}(g, sink.nodes = NULL)
}
\arguments{
\item{g}{graphNEL graph}

\item{sink.nodes}{to be documented}
}
\description{
Plots the pathway with KEGG layout based on x and y coordinates taken from KGML file.
}
