% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{plot_kegg_image_pathway}
\alias{plot_kegg_image_pathway}
\title{Plots the pathway and with colored nodes and labels}
\usage{
plot_kegg_image_pathway(
  pathway,
  no_color_mode = T,
  mapping_data_type = "signal",
  log_norm = TRUE,
  use_old_images = FALSE,
  plot_type = "kegg"
)
}
\arguments{
\item{pathway}{list object from kegg collection}

\item{no_color_mode}{when set to FALSE pathway nodes will be color coded with log expression FC values or log PSF values and color legend will be added to the pathway plot. Default value is TRUE}

\item{mapping_data_type}{type on node values to bi visualized. Possible values c("signal", "exp").}

\item{log_norm}{log transform PSF and expression values before color mapping. Default value is TRUE}

\item{use_old_images}{use_old_images use olde kegg images(for use with curated pathway collection).}

\item{plot_type}{network visualization type. Possible values c("kegg", "visnet"). When kegg option is used the bathway will be plotted over kegg png image. With visnet option function will plot interactive network with kegg layou.}
}
\description{
Plots the pathway and with colored nodes and labels
}
