% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PSF.R
\name{plot_PSF}
\alias{plot_PSF}
\title{Generate the plot for Forecasted Values}
\usage{
plot_PSF(data_in, predicted_data)
}
\arguments{
\item{data_in}{as Input data, in any format (data matrix data frame or vector). All variables should be numeric.}

\item{predicted_data}{as predicted data, in any format (data matrix data frame or vector). All variables should be numeric}
}
\value{
PSF_plot returns the plot of Predicted data
}
\description{
Takes Input data and Predicted data as input
}
\examples{
## Generate 100 random numbers within some limits
x <- sample(1:7, 100, replace = TRUE)
y <- sample(1:4, 20, replace = TRUE)
plot_PSF(x,y)
}

