% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{process.compounds}
\alias{process.compounds}
\title{Process protein compound interactions}
\usage{
process.compounds(g, edge.attrs)
}
\arguments{
\item{g}{The pathway graph of graphNEL class}

\item{edge.attrs}{The list of edge attributes in the pathway graph}
}
\description{
Process protein compound interactions
}
\details{
This function will turn gene -> gene interactions that occur via a compound node into gene -> compound -> gene interactions.
}
