% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{process.groupNode}
\alias{process.groupNode}
\title{Extend the group node to its component gene nodes}
\usage{
process.groupNode(groupNode, components, g, edge.attrs)
}
\arguments{
\item{groupNode}{The group node to be processed}

\item{components}{The vector of its component gene nodes}

\item{g}{The pathway graph of type GraphNEL}

\item{edge.attrs}{The list of edge attr}
}
\description{
Extend the group node to its component gene nodes
}
\details{
The all the edges to and from the group node are
}
