% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathnet.functions.R
\name{process.psf.results}
\alias{process.psf.results}
\title{Processs psf results output and calculates pvalues and some statitics}
\usage{
process.psf.results(psf.results)
}
\arguments{
\item{psf.results}{psf results output list}
}
\description{
Processs psf results output and calculates pvalues and some statitics
}
\keyword{internal}
