% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathnet.functions.R
\name{psf.from.env.entrez.fc}
\alias{psf.from.env.entrez.fc}
\title{Calculates pathway activity with PSF algorithm for provided kegg collection based on expression fold change data}
\usage{
psf.from.env.entrez.fc(
  entrez.fc,
  kegg.collection,
  split = TRUE,
  calculate.significance = T,
  bst.steps = 200,
  sum = FALSE,
  map_exp_data = TRUE
)
}
\arguments{
\item{entrez.fc}{expression fold change matrix with gene etrez id rownames}

\item{kegg.collection}{the list of kegg pathways generated by generate.kegg.collection.from.kgml or generate.kegg.collection function}

\item{split}{logical, if true then the incoming signal will be proportionally splitted among the edges}

\item{calculate.significance}{logical, if true then function will also calculate significance for the PSF values by shuffling all the network nodes in checking if the resulted PSF values were calculated by chance}

\item{bst.steps}{integer, the number of the interations for shuffling and recalculating PSF values for the significance analysis}

\item{sum}{logical, default value is FALSE. When set to true pathway activity will be caculated via addition, when set to false then activity willbe calculated via multiplication}

\item{map_exp_data}{logical, default value is TRUE. When set to false the expression data will not be mapped into pathway nodes and pathway node expression values will be used instead.}
}
\description{
Calculates pathway activity with PSF algorithm for provided kegg collection based on expression fold change data
}
