% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{redirectEdge}
\alias{redirectEdge}
\title{Redirect the edge to new source and target nodes}
\usage{
redirectEdge(prevSource, prevTarget, newSource, newTarget, g, edge.attrs)
}
\arguments{
\item{prevSource}{The original source node}

\item{prevTarget}{The original target node}

\item{newSource}{The new source node}

\item{newTarget}{The new target node}

\item{g}{The pathway graph of graphNEL class}

\item{edge.attrs}{The list of edge attributes in the graph}
}
\description{
Redirect the edge to new source and target nodes
}
\details{
A new edge between the new source and target nodes will be added to the graph, the previous edge will be removed, and its attributes will be transfered to the new edge.
}
