% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{remove.disconnected.nodes}
\alias{remove.disconnected.nodes}
\title{Remove nodes which do not have any interactions with other nodes}
\usage{
remove.disconnected.nodes(g)
}
\arguments{
\item{g}{graphNEL graph}
}
\description{
Remove nodes which do not have any interactions with other nodes
}
