% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{reverseEdge}
\alias{reverseEdge}
\title{Reverse edge direction}
\usage{
reverseEdge(snode, tnode, g, edge.attrs)
}
\arguments{
\item{snode}{parent node id}

\item{tnode}{child node id}

\item{g}{graphNEL grap}

\item{edge.attrs}{edge attributes}
}
\description{
Reverse edge direction
}
