% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_psf_functions.R
\name{run_psf_spatial_browser}
\alias{run_psf_spatial_browser}
\title{Launches PSF spatial browser Shiny application.}
\usage{
run_psf_spatial_browser(psf_saptial_results)
}
\arguments{
\item{psf_saptial_results}{list object generated by spatial_psf_analysis function.}
}
\description{
Launches PSF spatial browser Shiny application.
}
