% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{set.edge.impacts}
\alias{set.edge.impacts}
\title{Sets edge impacts for PSF analysis}
\usage{
set.edge.impacts(
  g,
  subtype1.attr = "subtype1",
  subtype2.attr = "subtype2",
  impact.attr.name = "impact"
)
}
\arguments{
\item{g}{GraphNEL object}

\item{subtype1.attr}{attribut name for subtype1}

\item{subtype2.attr}{attribute name for subtype2}

\item{impact.attr.name}{name of impact attribut}
}
\description{
Sets edge impacts for PSF analysis
}
