% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_psf_functions.R
\name{spatial_psf_analysis}
\alias{spatial_psf_analysis}
\title{Performs pathway activity analysis of Spatial transcriptomics data and subsequent clustering with Seurat clustering. Input data is a Seurat object.}
\usage{
spatial_psf_analysis(
  spatial_obj,
  pathway_collection,
  gene_symbol_to_entrez,
  nthreads = 1,
  return_only_shiny_vars = TRUE
)
}
\arguments{
\item{spatial_obj}{Seurat object of spatial transcriptomic dataset.}

\item{pathway_collection}{pathway collection list which will be used of PSF analysis}

\item{gene_symbol_to_entrez}{gene id conversion vector where names are the gene identifiers corresponding to count matrix row names and objects are entrez gene ids.}

\item{nthreads}{integer, the number of CPUs to use for calculation. Default value is 1.}

\item{return_only_shiny_vars}{when set tot TRUE list of variables will be return which can be used to lunch PSF spatial browser shiny app. Default value is TRUE.}
}
\description{
Performs pathway activity analysis of Spatial transcriptomics data and subsequent clustering with Seurat clustering. Input data is a Seurat object.
}
