% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_buffer_p.R
\name{st_buffer_parallel}
\alias{st_buffer_parallel}
\title{Combine sf and furrr to parallelize sf functions}
\usage{
st_buffer_parallel(
  x,
  rows_per_chunk = 1,
  .options = furrr::furrr_options(),
  ...
)
}
\arguments{
\item{x}{object of class sfg, sfc or sf}

\item{rows_per_chunk}{The data.frame will be split into chunks of this many
rows. Default is 1 (rowwise).}

\item{.options}{The \code{future} specific options to use with the workers. This
must be the result from a call to \code{\link[furrr:furrr_options]{furrr_options()}}.}

\item{...}{arguments passed on to `sf::st_buffer`}
}
\description{
Combine sf and furrr to parallelize sf functions
}
