% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.functions.R
\name{update_edge_weights}
\alias{update_edge_weights}
\title{Import edge weights extracted(further edited) via edge_data_frame_from_graph function}
\usage{
update_edge_weights(edge_weights, g)
}
\arguments{
\item{edge_weights}{vector with edge weigths with the order of edge_table generated by edge_data_frame_from_graph function}

\item{g}{graphNEL graph which will be updated}
}
\description{
Import edge weights extracted(further edited) via edge_data_frame_from_graph function
}
\author{
Siras Hakobyan
}
