% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psf_report_generator.R
\name{calc_psf_and_generate_report_from_collection}
\alias{calc_psf_and_generate_report_from_collection}
\title{Calculates psf for given kegg pathway based on expression matrix and generates pdf report with colored pathways and plots}
\usage{
calc_psf_and_generate_report_from_collection(
  kegg_collection,
  exp_matrix,
  file_name
)
}
\arguments{
\item{kegg_collection}{list of kegg pathways}

\item{exp_matrix}{expression fold change matrix with gene entrez id rownames}

\item{file_name}{name of the pdf report which will be generated}
}
\description{
Calculates psf for given kegg pathway based on expression matrix and generates pdf report with colored pathways and plots
}
