% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keggnet.functions.R
\name{map.gene.data}
\alias{map.gene.data}
\title{Map gene data onto a pathway graph}
\usage{
map.gene.data(g, gene.data)
}
\arguments{
\item{g}{Tha pathway graph of graphNEL class}

\item{gene.data}{A matrix data associated with the genes; rownames represent genes; a single gene-row may contain one or many data values;}
}
\value{
graph object with the column-wise averaged gene data kept in the nodedata attribute of the graph
}
\description{
Map gene data onto a pathway graph
}

