% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_interactive_app.R
\name{pathway_shiny_vis}
\alias{pathway_shiny_vis}
\title{Run KEGG interactive visualization App for provided pathway}
\usage{
pathway_shiny_vis(pathway)
}
\arguments{
\item{pathway}{list object}
}
\value{
Nothing is returned
}
\description{
Run KEGG interactive visualization App for provided pathway
}
\examples{
\dontrun{
    runShinyApp()
}

}
\concept{psf}
