% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_interactive_app.R
\name{run_psf_spatial_browser}
\alias{run_psf_spatial_browser}
\title{Run spatial PSF browser App}
\usage{
run_psf_spatial_browser(psf_saptial_results, port = 3838)
}
\arguments{
\item{psf_saptial_results}{list object generated by spatial_psf_analysis function.}

\item{port}{The TCP port that the application should listen on. Default value is 3838}
}
\value{
Nothing is returned
}
\description{
Run spatial PSF browser App
}
\examples{
\dontrun{
    runShinyApp()
}

}
\concept{psf}
