% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_interactive_app.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{Run KEGG interactive editor and visualization App}
\usage{
run_shiny_app(port = 3838)
}
\arguments{
\item{port}{the port number to use (Default: 3838)}
}
\value{
Nothing is returned
}
\description{
Run KEGG interactive editor and visualization App
}
\examples{
\dontrun{
    runShinyApp()
}

}
\concept{psf}
